added mocks test

+++ src/models/users.js
@@ -60,8 +60,29 @@ const userSchema = new mongoose.Schema(
 userSchema.virtual('tasks', {
 	ref: 'Task',
 	localField: '_id',
-	foreignField: 'user',
+	foreignField: 'owner',
 })
+
+userSchema.methods.toJSON = function () {
+	const user = this
+	const userObject = user.toObject()
+
+	delete userObject.password
+	delete userObject.tokens
+	delete userObject.avatar
+
+	return userObject
+}
+
+userSchema.methods.generateAuthToken = async function () {
+	const user = this
+	const token = jwt.sign({ _id: user._id.toString() }, process.env.JWT_SECRET)
+
+	user.tokens = user.tokens.concat({ token })
+	await user.save()
+
+	return token
+}
 
 userSchema.statics.findByCredentials = async (email, password) => {
 	const user = await User.findOne({ email })
@@ -78,26 +99,6 @@ userSchema.statics.findByCredentials = async (email, password) => {
 
 	return user
 }
-
-userSchema.methods.generateAuthToken = async function () {
-	const user = this
-	const token = jwt.sign({ _id: user._id.toString() }, process.env.JWT_SECRET)
-
-	user.tokens = user.tokens.concat({ token })
-	await user.save()
-
-	return token
-}
-
-userSchema.method.showProfile = async function () {
-	const user = this
-	const userObject = user.toObject()
-	delete userObject.password
-	delete userObject.tokens
-	delete userObject.avatar
-	return userObject
-}
-
 userSchema.pre('save', async function (next) {
 	const user = this
 
@@ -108,12 +109,13 @@ userSchema.pre('save', async function (next) {
 	next()
 })
 
+// Delete user tasks when user is removed
 userSchema.pre('remove', async function (next) {
 	const user = this
-	await Task.deleteMany({
-		owner: user._id,
-	})
+	await Task.deleteMany({ owner: user._id })
+	next()
 })
+
 const User = mongoose.model('User', userSchema)
 
 module.exports = User
